block: Add API for urgent request handling
This patch add support in block & elevator layers for handling
urgent requests. The decision if a request is urgent or not is taken
by the scheduler. Urgent request notification is passed to the
underlying
block device driver (eMMC for example). Block device driver may decide
to
interrupt the currently running low priority request to serve the new
urgent request. By doing so READ latency is greatly reduced in
read&write
collision scenarios.

Note that if the current scheduler doesn't implement the urgent request
mechanism, this code path is never activated.